# pkg/dashboard/swaggerdocs/swagger.yaml
...
basePath: /api
definitions:
...
  common.MapSlice:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
  common.Pod:
    properties:
      ip:
        type: string
      name:
        type: string
      namespace:
        type: string
      state:
        type: string
    type: object
  config.ChaosDashboardConfig:
    properties:
      cluster_mode:
        default: true
        description: ClusterScoped means control Chaos Object in cluster level(all
          namespace).
        type: boolean
      dns_server_create:
        default: false
        type: boolean
      enableFilterNamespace:
        default: false
        description: |-
          EnableFilterNamespace will filter namespace with annotation. Only the pods/containers in namespace
          annotated with `chaos-mesh.org/inject=enabled` will be injected.
        type: boolean
      listen_host:
        default: 0.0.0.0
        type: string
      listen_port:
        default: 2333
        type: integer
      security_mode:
        default: true
        description: If SecurityMode is set to true, then the dashboard requires token
          authentication.
        type: boolean
      target_namespace:
        description: |-
          TargetNamespace is the target namespace to injecting chaos.
          It only works with ClusterScoped is false.
        type: string
      version:
        type: string
    type: object
...
  core.SelectorInfo:
    properties:
      annotation_selectors:
        additionalProperties:
          type: string
        type: object
      expression_selectors:
        type: string
      field_selectors:
        additionalProperties:
          type: string
        type: object
      label_selectors:
        additionalProperties:
          type: string
        type: object
      namespaces:
        items:
          type: string
        type: array
      node_selectors:
        additionalProperties:
          type: string
        type: object
      nodes:
        items:
          type: string
        type: array
      phase_selectors:
        items:
          type: string
        type: array
      pods:
        additionalProperties:
          items:
            type: string
          type: array
        description: |-
          Pods is a map of string keys and a set values that used to select pods.
          The key defines the namespace which pods belong,
          and the each values is a set of pod names.
        type: object
    type: object
...
  utils.APIError:
    properties:
      code:
        type: integer
      full_text:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    name: GitHub Issues
    url: https://github.com/chaos-mesh/chaos-mesh/issues
  description: Swagger for Chaos Mesh Dashboard. If you encounter any problems with
    API, please click on the issues link below to report.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Chaos Mesh Dashboard API
  version: "2.0"
paths:
...
  /common/annotations:
    get:
      description: Get the annotations of the pods in the specified namespace from
        Kubernetes cluster.
      parameters:
      - description: The pod's namespace list, split by ,
        in: query
        name: podNamespaceList
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.MapSlice'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the annotations of the pods in the specified namespace from Kubernetes
        cluster.
      tags:
      - common
  /common/chaos-available-namespaces:
    get:
      description: Get all namespaces which could inject chaos(explosion scope) from
        Kubernetes cluster.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get all namespaces which could inject chaos(explosion scope) from Kubernetes
        cluster.
      tags:
      - common
  /common/config:
    get:
      description: Get the config of Dashboard.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.ChaosDashboardConfig'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the config of Dashboard.
      tags:
      - common
  /common/kinds:
    get:
      description: Get all chaos kinds from Kubernetes cluster.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get all chaos kinds from Kubernetes cluster.
      tags:
      - common
  /common/labels:
    get:
      description: Get the labels of the pods in the specified namespace from Kubernetes
        cluster.
      parameters:
      - description: The pod's namespace list, split by ,
        in: query
        name: podNamespaceList
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.MapSlice'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the labels of the pods in the specified namespace from Kubernetes
        cluster.
      tags:
      - common
  /common/namespaces:
    get:
      deprecated: true
      description: Get all from Kubernetes cluster.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get all namespaces from Kubernetes cluster.
      tags:
      - common
  /common/pods:
    post:
      description: Get pods from Kubernetes cluster.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.SelectorInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Pod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get pods from Kubernetes cluster.
      tags:
      - common
  /common/rbac-config:
    get:
      description: Get the rbac config according to the user's choice.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.MapSlice'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the rbac config according to the user's choice.
      tags:
      - common
...
swagger: "2.0"